refactor(facade): split incident responsibilities #185
