+++ c8volt/process/client_test.go
@@ -8,8 +8,6 @@ import (
 	"context"
 	"errors"
 	"log/slog"
-	"reflect"
-	"strings"
 	"sync/atomic"
 	"testing"
 	"time"
@@ -264,670 +262,6 @@ func TestClient_SearchProcessInstances_PreservesDerivedRelativeDayBoundsAsCanoni
 	}, 10, options.WithVerbose())
 
 	require.NoError(t, err)
-}
-
-// TestClient_SearchProcessInstanceIncidents_MapsDomainDetailsAndOptions verifies the facade keeps incident detail fields and call options intact.
-func TestClient_SearchProcessInstanceIncidents_MapsDomainDetailsAndOptions(t *testing.T) {
-	t.Parallel()
-
-	ctx := context.Background()
-	incAPI := stubIncidentAPI{
-		searchProcessInstanceIncidents: func(_ context.Context, key string, opts ...services.CallOption) ([]d.ProcessInstanceIncidentDetail, error) {
-			assert.Equal(t, "2251799813711967", key)
-			cfg := services.ApplyCallOptions(opts)
-			assert.True(t, cfg.Verbose)
-			assert.Equal(t, "all", cfg.IncidentState)
-			assert.Equal(t, "job_no_retries", cfg.IncidentErrorType)
-			assert.Equal(t, "No Retries", cfg.IncidentErrorMessage)
-			return []d.ProcessInstanceIncidentDetail{
-				{
-					IncidentKey:            "4503599627370497",
-					ProcessInstanceKey:     "2251799813711967",
-					TenantId:               "tenant-a",
-					State:                  "ACTIVE",
-					ErrorType:              "JOB_NO_RETRIES",
-					ErrorMessage:           "No retries left",
-					FlowNodeId:             "task-a",
-					FlowNodeInstanceKey:    "2251799813711999",
-					JobKey:                 "2251799813712000",
-					RootProcessInstanceKey: "2251799813711967",
-					ProcessDefinitionKey:   "2251799813685255",
-					ProcessDefinitionId:    "order-process",
-				},
-			}, nil
-		},
-	}
-
-	cli := New(&stubProcessDefinitionAPI{}, stubProcessInstanceAPI{}, incAPI, slog.Default())
-	got, err := cli.SearchProcessInstanceIncidents(ctx, "2251799813711967",
-		options.WithVerbose(),
-		options.WithIncidentState("all"),
-		options.WithIncidentErrorType("job_no_retries"),
-		options.WithIncidentErrorMessage("No Retries"),
-	)
-
-	require.NoError(t, err)
-	require.Equal(t, []ProcessInstanceIncidentDetail{
-		{
-			IncidentKey:            "4503599627370497",
-			ProcessInstanceKey:     "2251799813711967",
-			TenantId:               "tenant-a",
-			State:                  "ACTIVE",
-			ErrorType:              "JOB_NO_RETRIES",
-			ErrorMessage:           "No retries left",
-			FlowNodeId:             "task-a",
-			FlowNodeInstanceKey:    "2251799813711999",
-			JobKey:                 "2251799813712000",
-			RootProcessInstanceKey: "2251799813711967",
-			ProcessDefinitionKey:   "2251799813685255",
-			ProcessDefinitionId:    "order-process",
-		},
-	}, got)
-}
-
-func TestClient_GetIncidentAndSearchIncidentsMapServiceBoundary(t *testing.T) {
-	t.Parallel()
-
-	ctx := context.Background()
-	var calls []string
-	incAPI := stubIncidentAPI{
-		getIncident: func(_ context.Context, key string, opts ...services.CallOption) (d.ProcessInstanceIncidentDetail, error) {
-			calls = append(calls, "get:"+key)
-			cfg := services.ApplyCallOptions(opts)
-			assert.True(t, cfg.Verbose)
-			return d.ProcessInstanceIncidentDetail{
-				IncidentKey:        key,
-				ProcessInstanceKey: "pi-a",
-				State:              "ACTIVE",
-				ErrorType:          "JOB_NO_RETRIES",
-				ErrorMessage:       "no retries left",
-			}, nil
-		},
-		searchIncidents: func(_ context.Context, filter d.IncidentFilter, size int32, opts ...services.CallOption) ([]d.ProcessInstanceIncidentDetail, error) {
-			calls = append(calls, "search")
-			assert.Equal(t, d.IncidentFilter{
-				State:                  "resolved",
-				ErrorType:              "io_mapping_error",
-				ProcessInstanceKey:     "pi-a",
-				RootProcessInstanceKey: "root-pi-a",
-				ProcessDefinitionKey:   "pd-key",
-				ProcessDefinitionId:    "pd-id",
-				FlowNodeId:             "task-a",
-				FlowNodeInstanceKey:    "fni-a",
-			}, filter)
-			assert.Equal(t, int32(50), size)
-			assert.True(t, services.ApplyCallOptions(opts).Verbose)
-			return []d.ProcessInstanceIncidentDetail{{IncidentKey: "incident-b", ProcessInstanceKey: "pi-a"}}, nil
-		},
-	}
-
-	cli := New(&stubProcessDefinitionAPI{}, stubProcessInstanceAPI{}, incAPI, slog.Default())
-	gotIncident, err := cli.GetIncident(ctx, "incident-a", options.WithVerbose())
-	require.NoError(t, err)
-	gotSearch, err := cli.SearchIncidents(ctx, IncidentFilter{
-		State:                  "resolved",
-		ErrorType:              "io_mapping_error",
-		ProcessInstanceKey:     "pi-a",
-		RootProcessInstanceKey: "root-pi-a",
-		ProcessDefinitionKey:   "pd-key",
-		ProcessDefinitionId:    "pd-id",
-		FlowNodeId:             "task-a",
-		FlowNodeInstanceKey:    "fni-a",
-	}, 50, options.WithVerbose())
-
-	require.NoError(t, err)
-	require.Equal(t, []string{"get:incident-a", "search"}, calls)
-	require.Equal(t, "incident-a", gotIncident.IncidentKey)
-	require.Equal(t, int32(1), gotSearch.Total)
-	require.Equal(t, "incident-b", gotSearch.Items[0].IncidentKey)
-}
-
-func TestClient_SearchIncidentsWithMessageFilterPagesUntilEnoughLocalMatches(t *testing.T) {
-	t.Parallel()
-
-	ctx := context.Background()
-	var pages []d.IncidentPageRequest
-	incAPI := stubIncidentAPI{
-		searchIncidentsPage: func(_ context.Context, filter d.IncidentFilter, page d.IncidentPageRequest, opts ...services.CallOption) (d.IncidentPage, error) {
-			assert.Equal(t, d.IncidentFilter{State: "active", ErrorMessage: "intentional"}, filter)
-			assert.True(t, services.ApplyCallOptions(opts).Verbose)
-			pages = append(pages, page)
-			if len(pages) == 1 {
-				return d.IncidentPage{
-					Request:       page,
-					OverflowState: d.ProcessInstanceOverflowStateHasMore,
-					Items:         nil,
-				}, nil
-			}
-			return d.IncidentPage{
-				Request:       page,
-				OverflowState: d.ProcessInstanceOverflowStateNoMore,
-				Items: []d.ProcessInstanceIncidentDetail{
-					{IncidentKey: "match", ErrorMessage: "Intentional failure"},
-				},
-			}, nil
-		},
-	}
-
-	got, err := New(&stubProcessDefinitionAPI{}, stubProcessInstanceAPI{}, incAPI, slog.Default()).SearchIncidents(ctx, IncidentFilter{
-		State:        "active",
-		ErrorMessage: "intentional",
-	}, 1, options.WithVerbose())
-
-	require.NoError(t, err)
-	require.Equal(t, []d.IncidentPageRequest{{Size: 1}, {From: 1, Size: 1}}, pages)
-	require.Equal(t, int32(1), got.Total)
-	require.Equal(t, "match", got.Items[0].IncidentKey)
-}
-
-func TestClient_SearchIncidentsWithCreationTimeFilterPagesUntilEnoughLocalMatches(t *testing.T) {
-	t.Parallel()
-
-	ctx := context.Background()
-	var pages []d.IncidentPageRequest
-	incAPI := stubIncidentAPI{
-		searchIncidentsPage: func(_ context.Context, filter d.IncidentFilter, page d.IncidentPageRequest, opts ...services.CallOption) (d.IncidentPage, error) {
-			assert.Equal(t, d.IncidentFilter{State: "active", CreationTimeAfter: "2026-05-09T09:00:00Z"}, filter)
-			assert.True(t, services.ApplyCallOptions(opts).Verbose)
-			pages = append(pages, page)
-			if len(pages) == 1 {
-				return d.IncidentPage{
-					Request:       page,
-					OverflowState: d.ProcessInstanceOverflowStateHasMore,
-					Items:         nil,
-				}, nil
-			}
-			return d.IncidentPage{
-				Request:       page,
-				OverflowState: d.ProcessInstanceOverflowStateNoMore,
-				Items: []d.ProcessInstanceIncidentDetail{
-					{IncidentKey: "match", CreationTime: "2026-05-09T10:00:00Z"},
-				},
-			}, nil
-		},
-	}
-
-	got, err := New(&stubProcessDefinitionAPI{}, stubProcessInstanceAPI{}, incAPI, slog.Default()).SearchIncidents(ctx, IncidentFilter{
-		State:             "active",
-		CreationTimeAfter: "2026-05-09T09:00:00Z",
-	}, 1, options.WithVerbose())
-
-	require.NoError(t, err)
-	require.Equal(t, []d.IncidentPageRequest{{Size: 1}, {From: 1, Size: 1}}, pages)
-	require.Equal(t, int32(1), got.Total)
-	require.Equal(t, "match", got.Items[0].IncidentKey)
-}
-
-func TestClient_SearchIncidentsPageMapsMetadataAndUnsupportedErrors(t *testing.T) {
-	t.Parallel()
-
-	ctx := context.Background()
-	incAPI := stubIncidentAPI{
-		searchIncidentsPage: func(_ context.Context, filter d.IncidentFilter, page d.IncidentPageRequest, opts ...services.CallOption) (d.IncidentPage, error) {
-			assert.Equal(t, d.IncidentFilter{State: "active"}, filter)
-			assert.Equal(t, d.IncidentPageRequest{From: 10, Size: 5, After: "cursor-0"}, page)
-			assert.True(t, services.ApplyCallOptions(opts).Verbose)
-			return d.IncidentPage{
-				Request:       page,
-				OverflowState: d.ProcessInstanceOverflowStateHasMore,
-				ReportedTotal: &d.IncidentReportedTotal{
-					Count: 17,
-					Kind:  d.IncidentReportedTotalKindExact,
-				},
-				EndCursor: "cursor-1",
-				Items: []d.ProcessInstanceIncidentDetail{
-					{IncidentKey: "incident-a", ProcessInstanceKey: "pi-a"},
-				},
-			}, nil
-		},
-	}
-
-	cli := New(&stubProcessDefinitionAPI{}, stubProcessInstanceAPI{}, incAPI, slog.Default())
-	got, err := cli.SearchIncidentsPage(ctx, IncidentFilter{State: "active"}, IncidentPageRequest{From: 10, Size: 5, After: "cursor-0"}, options.WithVerbose())
-
-	require.NoError(t, err)
-	require.Equal(t, IncidentPageRequest{From: 10, Size: 5, After: "cursor-0"}, got.Request)
-	require.Equal(t, ProcessInstanceOverflowStateHasMore, got.OverflowState)
-	require.NotNil(t, got.ReportedTotal)
-	require.EqualValues(t, 17, got.ReportedTotal.Count)
-	require.Equal(t, IncidentReportedTotalKindExact, got.ReportedTotal.Kind)
-	require.Equal(t, "cursor-1", got.EndCursor)
-	require.Equal(t, "incident-a", got.Items[0].IncidentKey)
-
-	unsupportedAPI := stubIncidentAPI{
-		searchIncidentsPage: func(context.Context, d.IncidentFilter, d.IncidentPageRequest, ...services.CallOption) (d.IncidentPage, error) {
-			return d.IncidentPage{}, d.ErrUnsupported
-		},
-	}
-	_, err = New(&stubProcessDefinitionAPI{}, stubProcessInstanceAPI{}, unsupportedAPI, slog.Default()).SearchIncidentsPage(ctx, IncidentFilter{}, IncidentPageRequest{Size: 1})
-	require.Error(t, err)
-}
-
-func TestResolveIncidentWaitsForConfirmation(t *testing.T) {
-	t.Parallel()
-
-	ctx := context.Background()
-	var calls []string
-	incAPI := stubIncidentAPI{
-		getIncident: func(_ context.Context, key string, opts ...services.CallOption) (d.ProcessInstanceIncidentDetail, error) {
-			calls = append(calls, "get:"+key)
-			require.True(t, services.ApplyCallOptions(opts).Verbose)
-			return d.ProcessInstanceIncidentDetail{IncidentKey: key, ProcessInstanceKey: "2251799813685250", State: "ACTIVE"}, nil
-		},
-		resolveIncident: func(_ context.Context, key string, opts ...services.CallOption) (d.IncidentResolutionResponse, error) {
-			calls = append(calls, "resolve:"+key)
-			require.True(t, services.ApplyCallOptions(opts).Verbose)
-			return d.IncidentResolutionResponse{Key: key, Ok: true, StatusCode: 204, Status: "204 No Content"}, nil
-		},
-		waitForIncidentResolved: func(_ context.Context, key string, opts ...services.CallOption) (d.IncidentResolutionResponse, error) {
-			calls = append(calls, "wait:"+key)
-			require.True(t, services.ApplyCallOptions(opts).Verbose)
-			return d.IncidentResolutionResponse{Key: key, Ok: true, Status: "resolved"}, nil
-		},
-	}
-
-	cli := New(&stubProcessDefinitionAPI{}, stubProcessInstanceAPI{}, incAPI, slog.Default())
-	got, err := cli.ResolveIncident(ctx, "2251799813685249", options.WithVerbose())
-
-	require.NoError(t, err)
-	require.Equal(t, []string{"get:2251799813685249", "resolve:2251799813685249", "wait:2251799813685249"}, calls)
-	require.Equal(t, IncidentResolutionStatusConfirmed, got.Status)
-	require.True(t, got.MutationAccepted)
-	require.True(t, got.MutationSubmitted)
-	require.Equal(t, "resolved", got.ConfirmationStatus)
-}
-
-func TestResolveIncidentNoWaitSkipsConfirmation(t *testing.T) {
-	t.Parallel()
-
-	incAPI := stubIncidentAPI{
-		getIncident: func(_ context.Context, key string, opts ...services.CallOption) (d.ProcessInstanceIncidentDetail, error) {
-			require.Equal(t, "2251799813685249", key)
-			require.True(t, services.ApplyCallOptions(opts).NoWait)
-			return d.ProcessInstanceIncidentDetail{IncidentKey: key, ProcessInstanceKey: "2251799813685250", State: "ACTIVE"}, nil
-		},
-		resolveIncident: func(_ context.Context, key string, opts ...services.CallOption) (d.IncidentResolutionResponse, error) {
-			require.Equal(t, "2251799813685249", key)
-			require.True(t, services.ApplyCallOptions(opts).NoWait)
-			return d.IncidentResolutionResponse{Key: key, Ok: true, StatusCode: 204, Status: "204 No Content"}, nil
-		},
-	}
-
-	cli := New(&stubProcessDefinitionAPI{}, stubProcessInstanceAPI{}, incAPI, slog.Default())
-	got, err := cli.ResolveIncident(context.Background(), "2251799813685249", options.WithNoWait())
-
-	require.NoError(t, err)
-	require.Equal(t, IncidentResolutionStatusSubmitted, got.Status)
-	require.Equal(t, "skipped", got.ConfirmationStatus)
-}
-
-func TestResolveIncidentSkipsNonActiveIncidentBeforeMutation(t *testing.T) {
-	t.Parallel()
-
-	var mutated bool
-	incAPI := stubIncidentAPI{
-		getIncident: func(_ context.Context, key string, opts ...services.CallOption) (d.ProcessInstanceIncidentDetail, error) {
-			require.Equal(t, "2251799813685249", key)
-			return d.ProcessInstanceIncidentDetail{IncidentKey: key, ProcessInstanceKey: "2251799813685250", State: "RESOLVED"}, nil
-		},
-		resolveIncident: func(_ context.Context, key string, opts ...services.CallOption) (d.IncidentResolutionResponse, error) {
-			mutated = true
-			return d.IncidentResolutionResponse{}, nil
-		},
-	}
-
-	cli := New(&stubProcessDefinitionAPI{}, stubProcessInstanceAPI{}, incAPI, slog.Default())
-	got, err := cli.ResolveIncident(context.Background(), "2251799813685249")
-
-	require.NoError(t, err)
-	require.False(t, mutated)
-	require.Equal(t, IncidentResolutionStatusSkipped, got.Status)
-	require.Equal(t, "invalid_state", got.ConfirmationStatus)
-	require.Equal(t, "RESOLVED", got.IncidentState)
-	require.False(t, got.MutationSubmitted)
-}
-
-func TestResolveIncidentDryRunLoadsIncidentWithoutMutation(t *testing.T) {
-	t.Parallel()
-
-	incAPI := stubIncidentAPI{
-		getIncident: func(_ context.Context, key string, opts ...services.CallOption) (d.ProcessInstanceIncidentDetail, error) {
-			require.Equal(t, "2251799813685249", key)
-			require.True(t, services.ApplyCallOptions(opts).DryRun)
-			return d.ProcessInstanceIncidentDetail{
-				IncidentKey:        key,
-				ProcessInstanceKey: "2251799813685250",
-				State:              "ACTIVE",
-			}, nil
-		},
-	}
-
-	cli := New(&stubProcessDefinitionAPI{}, stubProcessInstanceAPI{}, incAPI, slog.Default())
-	got, err := cli.ResolveIncident(context.Background(), "2251799813685249", options.WithDryRun())
-
-	require.NoError(t, err)
-	require.Equal(t, IncidentResolutionStatusPlanned, got.Status)
-	require.True(t, got.DryRun)
-	require.False(t, got.MutationSubmitted)
-	require.True(t, got.WouldResolve)
-	require.Equal(t, "2251799813685250", got.ProcessInstanceKey)
-	require.NotNil(t, got.Incident)
-}
-
-func TestResolveIncidentUnsupportedMapsFailureBeforeWait(t *testing.T) {
-	t.Parallel()
-
-	incAPI := stubIncidentAPI{
-		getIncident: func(_ context.Context, key string, opts ...services.CallOption) (d.ProcessInstanceIncidentDetail, error) {
-			require.Equal(t, "2251799813685249", key)
-			return d.ProcessInstanceIncidentDetail{IncidentKey: key, ProcessInstanceKey: "2251799813685250", State: "ACTIVE"}, nil
-		},
-		resolveIncident: func(_ context.Context, key string, opts ...services.CallOption) (d.IncidentResolutionResponse, error) {
-			require.Equal(t, "2251799813685249", key)
-			return d.IncidentResolutionResponse{Key: key, Ok: false, Status: "unsupported"}, d.ErrUnsupported
-		},
-	}
-
-	cli := New(&stubProcessDefinitionAPI{}, stubProcessInstanceAPI{}, incAPI, slog.Default())
-	got, err := cli.ResolveIncident(context.Background(), "2251799813685249")
-
-	require.Error(t, err)
-	require.Equal(t, IncidentResolutionStatusMutationFailed, got.Status)
-	require.False(t, got.MutationSubmitted)
-	require.Contains(t, got.Error, "unsupported")
-}
-
-func TestResolveProcessInstanceIncidentsDiscoversAndWaitsInitialActiveSet(t *testing.T) {
-	t.Parallel()
-
-	var resolved []string
-	var waited []string
-	incAPI := stubIncidentAPI{
-		searchProcessInstanceIncidents: func(_ context.Context, key string, opts ...services.CallOption) ([]d.ProcessInstanceIncidentDetail, error) {
-			require.Equal(t, "2251799813685250", key)
-			require.True(t, services.ApplyCallOptions(opts).Verbose)
-			return []d.ProcessInstanceIncidentDetail{
-				{IncidentKey: "incident-a", ProcessInstanceKey: key, State: "ACTIVE"},
-				{IncidentKey: "incident-b", ProcessInstanceKey: key, State: "RESOLVED"},
-				{IncidentKey: "wrong-owner", ProcessInstanceKey: "other", State: "ACTIVE"},
-			}, nil
-		},
-		resolveIncident: func(_ context.Context, key string, opts ...services.CallOption) (d.IncidentResolutionResponse, error) {
-			resolved = append(resolved, key)
-			require.True(t, services.ApplyCallOptions(opts).Verbose)
-			return d.IncidentResolutionResponse{Key: key, Ok: true, StatusCode: 204, Status: "204 No Content"}, nil
-		},
-		waitForPIIncidentsResolved: func(_ context.Context, key string, incidentKeys []string, opts ...services.CallOption) (d.IncidentResolutionResponse, error) {
-			require.Equal(t, "2251799813685250", key)
-			waited = append(waited, incidentKeys...)
-			require.True(t, services.ApplyCallOptions(opts).Verbose)
-			return d.IncidentResolutionResponse{Key: key, Ok: true, Status: "resolved"}, nil
-		},
-	}
-
-	cli := New(&stubProcessDefinitionAPI{}, stubProcessInstanceAPI{}, incAPI, slog.Default())
-	got, err := cli.ResolveProcessInstanceIncidents(context.Background(), "2251799813685250", options.WithVerbose())
-
-	require.NoError(t, err)
-	require.Equal(t, []string{"incident-a"}, resolved)
-	require.Equal(t, []string{"incident-a"}, waited)
-	require.Equal(t, ProcessInstanceResolutionStatusConfirmed, got.Status)
-	require.Equal(t, []string{"incident-a"}, got.AttemptedIncidentKeys)
-	require.Equal(t, []string{"incident-a"}, got.ResolvedIncidentKeys)
-	require.True(t, got.MutationSubmitted)
-}
-
-func TestResolveProcessInstanceIncidentsDryRunPlansNoMutation(t *testing.T) {
-	t.Parallel()
-
-	incAPI := stubIncidentAPI{
-		searchProcessInstanceIncidents: func(_ context.Context, key string, opts ...services.CallOption) ([]d.ProcessInstanceIncidentDetail, error) {
-			require.True(t, services.ApplyCallOptions(opts).DryRun)
-			return []d.ProcessInstanceIncidentDetail{
-				{IncidentKey: "incident-a", ProcessInstanceKey: key, State: "ACTIVE"},
-			}, nil
-		},
-	}
-
-	cli := New(&stubProcessDefinitionAPI{}, stubProcessInstanceAPI{}, incAPI, slog.Default())
-	got, err := cli.ResolveProcessInstanceIncidents(context.Background(), "2251799813685250", options.WithDryRun())
-
-	require.NoError(t, err)
-	require.Equal(t, ProcessInstanceResolutionStatusPlanned, got.Status)
-	require.True(t, got.DryRun)
-	require.False(t, got.MutationSubmitted)
-	require.Equal(t, []string{"incident-a"}, got.AttemptedIncidentKeys)
-}
-
-func TestResolveProcessInstanceIncidentsReportsNoActiveIncidentsAsSkipped(t *testing.T) {
-	t.Parallel()
-
-	incAPI := stubIncidentAPI{
-		searchProcessInstanceIncidents: func(_ context.Context, key string, opts ...services.CallOption) ([]d.ProcessInstanceIncidentDetail, error) {
-			return []d.ProcessInstanceIncidentDetail{
-				{IncidentKey: "incident-a", ProcessInstanceKey: key, State: "RESOLVED"},
-			}, nil
-		},
-	}
-
-	cli := New(&stubProcessDefinitionAPI{}, stubProcessInstanceAPI{}, incAPI, slog.Default())
-	got, err := cli.ResolveProcessInstanceIncidents(context.Background(), "2251799813685250")
-
-	require.NoError(t, err)
-	require.Equal(t, ProcessInstanceResolutionStatusSkipped, got.Status)
-	require.Equal(t, "no_active_incidents", got.ConfirmationStatus)
-	require.False(t, got.MutationSubmitted)
-}
-
-func TestResolveProcessInstanceIncidentsReportsPartialMutationFailures(t *testing.T) {
-	t.Parallel()
-
-	incAPI := stubIncidentAPI{
-		searchProcessInstanceIncidents: func(_ context.Context, key string, opts ...services.CallOption) ([]d.ProcessInstanceIncidentDetail, error) {
-			return []d.ProcessInstanceIncidentDetail{
-				{IncidentKey: "incident-a", ProcessInstanceKey: key, State: "ACTIVE"},
-				{IncidentKey: "incident-b", ProcessInstanceKey: key, State: "ACTIVE"},
-			}, nil
-		},
-		resolveIncident: func(_ context.Context, key string, opts ...services.CallOption) (d.IncidentResolutionResponse, error) {
-			require.Contains(t, []string{"incident-a", "incident-b"}, key)
-			if key == "incident-a" {
-				return d.IncidentResolutionResponse{Key: key, Ok: true, StatusCode: 204, Status: "204 No Content"}, nil
-			}
-			return d.IncidentResolutionResponse{Key: key, Ok: false, Status: "500 Internal Server Error"}, errors.New("mutation rejected")
-		},
-	}
-
-	cli := New(&stubProcessDefinitionAPI{}, stubProcessInstanceAPI{}, incAPI, slog.Default())
-	got, err := cli.ResolveProcessInstanceIncidents(context.Background(), "2251799813685250", options.WithNoWait())
-
-	require.Error(t, err)
-	require.Equal(t, ProcessInstanceResolutionStatusPartialFailed, got.Status)
-	require.Equal(t, []string{"incident-a"}, got.ResolvedIncidentKeys)
-	require.Equal(t, []string{"incident-b"}, got.FailedIncidentKeys)
-	require.True(t, got.MutationSubmitted)
-}
-
-func TestResolveIncidentsBulkWorkersNoWorkerLimitAndPartialFailureTotals(t *testing.T) {
-	t.Parallel()
-
-	var active int32
-	var maxActive int32
-	var calls int32
-	incAPI := stubIncidentAPI{
-		getIncident: func(_ context.Context, key string, opts ...services.CallOption) (d.ProcessInstanceIncidentDetail, error) {
-			return d.ProcessInstanceIncidentDetail{IncidentKey: key, State: "ACTIVE"}, nil
-		},
-		resolveIncident: func(_ context.Context, key string, opts ...services.CallOption) (d.IncidentResolutionResponse, error) {
-			cfg := services.ApplyCallOptions(opts)
-			if !cfg.NoWait {
-				return d.IncidentResolutionResponse{}, errors.New("expected no-wait call option")
-			}
-			if !cfg.NoWorkerLimit {
-				return d.IncidentResolutionResponse{}, errors.New("expected no-worker-limit call option")
-			}
-			current := atomic.AddInt32(&active, 1)
-			for {
-				previous := atomic.LoadInt32(&maxActive)
-				if current <= previous || atomic.CompareAndSwapInt32(&maxActive, previous, current) {
-					break
-				}
-			}
-			time.Sleep(10 * time.Millisecond)
-			atomic.AddInt32(&active, -1)
-			atomic.AddInt32(&calls, 1)
-			if key == "incident-b" {
-				return d.IncidentResolutionResponse{Key: key, Ok: false, StatusCode: 500, Status: "500 Internal Server Error"}, errors.New("mutation rejected")
-			}
-			return d.IncidentResolutionResponse{Key: key, Ok: true, StatusCode: 204, Status: "204 No Content"}, nil
-		},
-	}
-
-	cli := New(&stubProcessDefinitionAPI{}, stubProcessInstanceAPI{}, incAPI, slog.Default())
-	got, err := cli.ResolveIncidents(context.Background(),
-		typex.Keys{"incident-a", "incident-c", "incident-b", "incident-a"},
-		2,
-		options.WithNoWait(),
-		options.WithNoWorkerLimit(),
-	)
-
-	require.Error(t, err)
-	require.Len(t, got.Items, 3)
-	require.Equal(t, 3, got.Total)
-	require.Equal(t, 2, got.Submitted)
-	require.Equal(t, 1, got.Failed)
-	require.True(t, got.MutationSubmitted)
-	require.Equal(t, int32(3), atomic.LoadInt32(&calls))
-	require.Equal(t, int32(2), atomic.LoadInt32(&maxActive))
-}
-
-func TestResolveIncidentsBulkFailFastStopsSchedulingAfterFirstFailure(t *testing.T) {
-	t.Parallel()
-
-	var calls int32
-	incAPI := stubIncidentAPI{
-		getIncident: func(_ context.Context, key string, opts ...services.CallOption) (d.ProcessInstanceIncidentDetail, error) {
-			require.Equal(t, "incident-fail", key)
-			require.True(t, services.ApplyCallOptions(opts).FailFast)
-			return d.ProcessInstanceIncidentDetail{IncidentKey: key, State: "ACTIVE"}, nil
-		},
-		resolveIncident: func(_ context.Context, key string, opts ...services.CallOption) (d.IncidentResolutionResponse, error) {
-			require.Equal(t, "incident-fail", key)
-			require.True(t, services.ApplyCallOptions(opts).FailFast)
-			atomic.AddInt32(&calls, 1)
-			return d.IncidentResolutionResponse{Key: key, Ok: false, StatusCode: 500, Status: "500 Internal Server Error"}, errors.New("mutation rejected")
-		},
-	}
-
-	cli := New(&stubProcessDefinitionAPI{}, stubProcessInstanceAPI{}, incAPI, slog.Default())
-	got, err := cli.ResolveIncidents(context.Background(),
-		typex.Keys{"incident-fail", "incident-skip"},
-		1,
-		options.WithFailFast(),
-		options.WithNoWait(),
-	)
-
-	require.Error(t, err)
-	require.Equal(t, int32(1), atomic.LoadInt32(&calls))
-	require.Len(t, got.Items, 1)
-	require.Equal(t, 1, got.Total)
-	require.Equal(t, 1, got.Failed)
-	require.Equal(t, "incident-fail", got.Items[0].IncidentKey)
-}
-
-func TestResolveProcessInstancesIncidentsBulkReportsPartialFailureTotals(t *testing.T) {
-	t.Parallel()
-
-	incAPI := stubIncidentAPI{
-		searchProcessInstanceIncidents: func(_ context.Context, key string, opts ...services.CallOption) ([]d.ProcessInstanceIncidentDetail, error) {
-			require.Equal(t, "pi-a", key)
-			require.True(t, services.ApplyCallOptions(opts).NoWait)
-			return []d.ProcessInstanceIncidentDetail{
-				{IncidentKey: "incident-a", ProcessInstanceKey: key, State: "ACTIVE"},
-				{IncidentKey: "incident-b", ProcessInstanceKey: key, State: "ACTIVE"},
-			}, nil
-		},
-		resolveIncident: func(_ context.Context, key string, opts ...services.CallOption) (d.IncidentResolutionResponse, error) {
-			require.True(t, services.ApplyCallOptions(opts).NoWait)
-			require.Contains(t, []string{"incident-a", "incident-b"}, key)
-			return d.IncidentResolutionResponse{Key: key, Ok: false, StatusCode: 500, Status: "500 Internal Server Error"}, errors.New("mutation rejected")
-		},
-	}
-
-	cli := New(&stubProcessDefinitionAPI{}, stubProcessInstanceAPI{}, incAPI, slog.Default())
-	got, err := cli.ResolveProcessInstancesIncidents(context.Background(),
-		typex.Keys{"pi-a"},
-		1,
-		options.WithNoWait(),
-	)
-
-	require.Error(t, err)
-	require.Equal(t, 1, got.Total)
-	require.Equal(t, 1, got.Failed)
-	require.False(t, got.MutationSubmitted)
-	require.Len(t, got.Items, 1)
-	require.Equal(t, ProcessInstanceResolutionStatusFailed, got.Items[0].Status)
-	require.Empty(t, got.Items[0].ResolvedIncidentKeys)
-	require.Equal(t, []string{"incident-a", "incident-b"}, got.Items[0].FailedIncidentKeys)
-}
-
-func TestResolveProcessInstancesIncidentsBulkFailFastStopsSchedulingAfterFirstLookupFailure(t *testing.T) {
-	t.Parallel()
-
-	var searches int32
-	incAPI := stubIncidentAPI{
-		searchProcessInstanceIncidents: func(_ context.Context, key string, opts ...services.CallOption) ([]d.ProcessInstanceIncidentDetail, error) {
-			require.Equal(t, "pi-fail", key)
-			cfg := services.ApplyCallOptions(opts)
-			require.True(t, cfg.FailFast)
-			require.True(t, cfg.NoWorkerLimit)
-			atomic.AddInt32(&searches, 1)
-			return nil, errors.New("lookup failed")
-		},
-	}
-
-	cli := New(&stubProcessDefinitionAPI{}, stubProcessInstanceAPI{}, incAPI, slog.Default())
-	got, err := cli.ResolveProcessInstancesIncidents(context.Background(),
-		typex.Keys{"pi-fail", "pi-skip"},
-		1,
-		options.WithFailFast(),
-		options.WithNoWorkerLimit(),
-	)
-
-	require.Error(t, err)
-	require.Equal(t, int32(1), atomic.LoadInt32(&searches))
-	require.Len(t, got.Items, 1)
-	require.Equal(t, 1, got.Total)
-	require.Equal(t, 1, got.Failed)
-	require.Equal(t, "pi-fail", got.Items[0].ProcessInstanceKey)
-}
-
-func TestResolveProcessInstanceIncidentsKeepsIncidentBoundaryOutOfProcessInstanceService(t *testing.T) {
-	t.Parallel()
-
-	apiType := reflect.TypeOf((*pisvc.API)(nil)).Elem()
-	for i := 0; i < apiType.NumMethod(); i++ {
-		method := apiType.Method(i)
-		require.NotContains(t, strings.ToLower(method.Name), "incident")
-	}
-
-	var lookedUp bool
-	incAPI := stubIncidentAPI{
-		searchProcessInstanceIncidents: func(_ context.Context, key string, opts ...services.CallOption) ([]d.ProcessInstanceIncidentDetail, error) {
-			lookedUp = true
-			require.Equal(t, "2251799813685250", key)
-			return []d.ProcessInstanceIncidentDetail{
-				{IncidentKey: "2251799813685249", ProcessInstanceKey: key, State: "RESOLVED"},
-			}, nil
-		},
-	}
-
-	cli := New(&stubProcessDefinitionAPI{}, stubProcessInstanceAPI{}, incAPI, slog.Default())
-	got, err := cli.ResolveProcessInstanceIncidents(context.Background(), "2251799813685250")
-
-	require.NoError(t, err)
-	require.True(t, lookedUp)
-	require.Equal(t, ProcessInstanceResolutionStatusSkipped, got.Status)
 }
 
 // TestClient_EnrichProcessInstancesWithIncidents_PreservesOrderAndPerKeyAssociation prevents incident details from leaking across keyed results.
