+++ c8volt/process/model.go
@@ -7,6 +7,7 @@ import (
 	"fmt"
 
 	ferr "github.com/grafvonb/c8volt/c8volt/ferrors"
+	"github.com/grafvonb/c8volt/c8volt/incident"
 	"github.com/grafvonb/c8volt/toolx"
 )
 
@@ -75,176 +76,7 @@ type ProcessInstance struct {
 	Variables                 map[string]any `json:"variables,omitempty"`
 }
 
-type ProcessInstanceIncidentDetail struct {
-	IncidentKey            string `json:"incidentKey,omitempty"`
-	CreationTime           string `json:"creationTime,omitempty"`
-	ProcessInstanceKey     string `json:"processInstanceKey"`
-	TenantId               string `json:"tenantId,omitempty"`
-	State                  string `json:"state,omitempty"`
-	ErrorType              string `json:"errorType,omitempty"`
-	ErrorMessage           string `json:"errorMessage"`
-	FlowNodeId             string `json:"flowNodeId,omitempty"`
-	FlowNodeInstanceKey    string `json:"flowNodeInstanceKey,omitempty"`
-	JobKey                 string `json:"jobKey,omitempty"`
-	RootProcessInstanceKey string `json:"rootProcessInstanceKey,omitempty"`
-	ProcessDefinitionKey   string `json:"processDefinitionKey,omitempty"`
-	ProcessDefinitionId    string `json:"processDefinitionId,omitempty"`
-}
-
-type IncidentFilter struct {
-	State                  string `json:"state,omitempty"`
-	ErrorType              string `json:"errorType,omitempty"`
-	ErrorMessage           string `json:"errorMessage,omitempty"`
-	ProcessInstanceKey     string `json:"processInstanceKey,omitempty"`
-	RootProcessInstanceKey string `json:"rootProcessInstanceKey,omitempty"`
-	ProcessDefinitionKey   string `json:"processDefinitionKey,omitempty"`
-	ProcessDefinitionId    string `json:"processDefinitionId,omitempty"`
-	FlowNodeId             string `json:"flowNodeId,omitempty"`
-	FlowNodeInstanceKey    string `json:"flowNodeInstanceKey,omitempty"`
-	CreationTimeAfter      string `json:"creationTimeAfter,omitempty"`
-	CreationTimeBefore     string `json:"creationTimeBefore,omitempty"`
-}
-
-type Incidents struct {
-	Total int32                           `json:"total,omitempty"`
-	Items []ProcessInstanceIncidentDetail `json:"items,omitempty"`
-}
-
-type IncidentPageRequest struct {
-	From  int32  `json:"from,omitempty"`
-	Size  int32  `json:"size,omitempty"`
-	After string `json:"after,omitempty"`
-}
-
-type IncidentReportedTotalKind string
-
-const (
-	IncidentReportedTotalKindExact      IncidentReportedTotalKind = "exact"
-	IncidentReportedTotalKindLowerBound IncidentReportedTotalKind = "lower_bound"
-)
-
-type IncidentReportedTotal struct {
-	Count int64                     `json:"count,omitempty"`
-	Kind  IncidentReportedTotalKind `json:"kind,omitempty"`
-}
-
-type IncidentPage struct {
-	Request       IncidentPageRequest             `json:"request,omitempty"`
-	OverflowState ProcessInstanceOverflowState    `json:"overflowState,omitempty"`
-	ReportedTotal *IncidentReportedTotal          `json:"reportedTotal,omitempty"`
-	EndCursor     string                          `json:"endCursor,omitempty"`
-	Items         []ProcessInstanceIncidentDetail `json:"items,omitempty"`
-}
-
-type ResolutionOperation string
-
-const (
-	ResolutionOperationIncident        ResolutionOperation = "resolveIncident"
-	ResolutionOperationProcessInstance ResolutionOperation = "resolveProcessInstance"
-)
-
-type IncidentResolutionStatus string
-
-const (
-	IncidentResolutionStatusPlanned            IncidentResolutionStatus = "planned"
-	IncidentResolutionStatusSubmitted          IncidentResolutionStatus = "submitted"
-	IncidentResolutionStatusConfirmed          IncidentResolutionStatus = "confirmed"
-	IncidentResolutionStatusSkipped            IncidentResolutionStatus = "skipped"
-	IncidentResolutionStatusMutationFailed     IncidentResolutionStatus = "mutation_failed"
-	IncidentResolutionStatusConfirmationFailed IncidentResolutionStatus = "confirmation_failed"
-)
-
-type IncidentResolutionResult struct {
-	IncidentKey        string                         `json:"incidentKey"`
-	ProcessInstanceKey string                         `json:"processInstanceKey,omitempty"`
-	MutationAccepted   bool                           `json:"mutationAccepted"`
-	Status             IncidentResolutionStatus       `json:"status"`
-	ConfirmationStatus string                         `json:"confirmationStatus,omitempty"`
-	StatusCode         int                            `json:"statusCode,omitempty"`
-	Message            string                         `json:"message,omitempty"`
-	Error              string                         `json:"error,omitempty"`
-	DryRun             bool                           `json:"dryRun"`
-	MutationSubmitted  bool                           `json:"mutationSubmitted"`
-	WouldResolve       bool                           `json:"wouldResolve,omitempty"`
-	IncidentState      string                         `json:"incidentState,omitempty"`
-	Incident           *ProcessInstanceIncidentDetail `json:"incident,omitempty"`
-}
-
-func (r IncidentResolutionResult) OK() bool {
-	return r.Status != IncidentResolutionStatusMutationFailed && r.Status != IncidentResolutionStatusConfirmationFailed
-}
-
-type IncidentResolutionResults struct {
-	Operation         ResolutionOperation        `json:"operation,omitempty"`
-	Items             []IncidentResolutionResult `json:"items,omitempty"`
-	Total             int                        `json:"total"`
-	Submitted         int                        `json:"submitted"`
-	Confirmed         int                        `json:"confirmed"`
-	Skipped           int                        `json:"skipped"`
-	Failed            int                        `json:"failed"`
-	DryRun            bool                       `json:"dryRun"`
-	MutationSubmitted bool                       `json:"mutationSubmitted"`
-}
-
-func (r IncidentResolutionResults) Totals() (total int, oks int, noks int) {
-	return TotalsOf(r.Items)
-}
-
-type ProcessInstanceResolutionStatus string
-
-const (
-	ProcessInstanceResolutionStatusPlanned       ProcessInstanceResolutionStatus = "planned"
-	ProcessInstanceResolutionStatusSubmitted     ProcessInstanceResolutionStatus = "submitted"
-	ProcessInstanceResolutionStatusConfirmed     ProcessInstanceResolutionStatus = "confirmed"
-	ProcessInstanceResolutionStatusSkipped       ProcessInstanceResolutionStatus = "skipped"
-	ProcessInstanceResolutionStatusPartialFailed ProcessInstanceResolutionStatus = "partial_failed"
-	ProcessInstanceResolutionStatusFailed        ProcessInstanceResolutionStatus = "failed"
-)
-
-type ProcessInstanceResolutionResult struct {
-	ProcessInstanceKey    string                          `json:"processInstanceKey"`
-	AttemptedIncidentKeys []string                        `json:"attemptedIncidentKeys,omitempty"`
-	ResolvedIncidentKeys  []string                        `json:"resolvedIncidentKeys,omitempty"`
-	SkippedIncidentKeys   []string                        `json:"skippedIncidentKeys,omitempty"`
-	FailedIncidentKeys    []string                        `json:"failedIncidentKeys,omitempty"`
-	ConfirmationStatus    string                          `json:"confirmationStatus,omitempty"`
-	Status                ProcessInstanceResolutionStatus `json:"status"`
-	Error                 string                          `json:"error,omitempty"`
-	DryRun                bool                            `json:"dryRun"`
-	MutationSubmitted     bool                            `json:"mutationSubmitted"`
-	Incidents             []ProcessInstanceIncidentDetail `json:"incidents,omitempty"`
-}
-
-func (r ProcessInstanceResolutionResult) OK() bool {
-	return r.Status != ProcessInstanceResolutionStatusFailed && r.Status != ProcessInstanceResolutionStatusPartialFailed
-}
-
-type ProcessInstanceResolutionResults struct {
-	Operation         ResolutionOperation               `json:"operation,omitempty"`
-	Items             []ProcessInstanceResolutionResult `json:"items,omitempty"`
-	Total             int                               `json:"total"`
-	Submitted         int                               `json:"submitted"`
-	Confirmed         int                               `json:"confirmed"`
-	Skipped           int                               `json:"skipped"`
-	Failed            int                               `json:"failed"`
-	DryRun            bool                              `json:"dryRun"`
-	MutationSubmitted bool                              `json:"mutationSubmitted"`
-}
-
-func (r ProcessInstanceResolutionResults) Totals() (total int, oks int, noks int) {
-	return TotalsOf(r.Items)
-}
-
-type ResolutionPlan struct {
-	Operation                ResolutionOperation `json:"operation"`
-	RequestedKeys            []string            `json:"requestedKeys,omitempty"`
-	DryRun                   bool                `json:"dryRun"`
-	MutationSubmitted        bool                `json:"mutationSubmitted"`
-	Items                    any                 `json:"items,omitempty"`
-	WouldResolveIncidentKeys []string            `json:"wouldResolveIncidentKeys,omitempty"`
-	SkippedIncidentKeys      []string            `json:"skippedIncidentKeys,omitempty"`
-	Errors                   []string            `json:"errors,omitempty"`
-}
+type ProcessInstanceIncidentDetail = incident.ProcessInstanceIncidentDetail
 
 type ProcessInstanceVariable struct {
 	Name               string `json:"name"`
