+++ internal/domain/report.go
@@ -1,0 +1,15 @@
+// SPDX-FileCopyrightText: 2026 Adam Bogdan Boczek
+// SPDX-License-Identifier: GPL-3.0-or-later
+
+package domain
+
+type Reporter struct {
+	Key        string
+	Ok         bool
+	StatusCode int
+	Status     string
+}
+
+type Reporters struct {
+	Items []Reporter
+}
