+++ internal/services/incident/resolve.go
@@ -11,62 +11,7 @@ import (
 	"github.com/grafvonb/c8volt/internal/services"
 	"github.com/grafvonb/c8volt/internal/services/incident/waiter"
 	"github.com/grafvonb/c8volt/toolx"
-	"github.com/grafvonb/c8volt/toolx/pool"
-	"github.com/grafvonb/c8volt/typex"
 )
-
-func GetIncidents(ctx context.Context, api API, keys typex.Keys, wantedWorkers int, opts ...services.CallOption) ([]d.ProcessInstanceIncidentDetail, error) {
-	cfg := services.ApplyCallOptions(opts)
-	ukeys := keys.Unique()
-	workers := toolx.DetermineNoOfWorkers(len(ukeys), wantedWorkers, cfg.NoWorkerLimit)
-	return pool.ExecuteSlice[string, d.ProcessInstanceIncidentDetail](ctx, ukeys, workers, cfg.FailFast, func(ctx context.Context, key string, _ int) (d.ProcessInstanceIncidentDetail, error) {
-		return api.GetIncident(ctx, key, opts...)
-	})
-}
-
-func SearchIncidents(ctx context.Context, api API, filter d.IncidentFilter, size int32, opts ...services.CallOption) ([]d.ProcessInstanceIncidentDetail, error) {
-	if incidentSearchNeedsPagedLocalFiltering(filter) {
-		return searchIncidentPagesUntilLimit(ctx, api, filter, size, opts...)
-	}
-	return api.SearchIncidents(ctx, filter, size, opts...)
-}
-
-func searchIncidentPagesUntilLimit(ctx context.Context, api API, filter d.IncidentFilter, size int32, opts ...services.CallOption) ([]d.ProcessInstanceIncidentDetail, error) {
-	if size <= 0 {
-		return nil, nil
-	}
-	req := d.IncidentPageRequest{Size: size}
-	out := make([]d.ProcessInstanceIncidentDetail, 0, size)
-	for {
-		page, err := api.SearchIncidentsPage(ctx, filter, req, opts...)
-		if err != nil {
-			return nil, err
-		}
-		for _, item := range page.Items {
-			if int32(len(out)) >= size {
-				return out, nil
-			}
-			out = append(out, item)
-		}
-		if page.OverflowState == d.ProcessInstanceOverflowStateNoMore {
-			return out, nil
-		}
-		req = nextIncidentPageRequest(req, page)
-	}
-}
-
-func incidentSearchNeedsPagedLocalFiltering(filter d.IncidentFilter) bool {
-	return filter.ErrorMessage != "" ||
-		filter.CreationTimeAfter != "" ||
-		filter.CreationTimeBefore != ""
-}
-
-func nextIncidentPageRequest(current d.IncidentPageRequest, page d.IncidentPage) d.IncidentPageRequest {
-	if page.EndCursor != "" {
-		return d.IncidentPageRequest{Size: current.Size, After: page.EndCursor}
-	}
-	return d.IncidentPageRequest{From: current.From + current.Size, Size: current.Size}
-}
 
 func ResolveIncidentOperation(ctx context.Context, api API, key string, opts ...services.CallOption) (d.IncidentResolutionResult, error) {
 	cfg := services.ApplyCallOptions(opts)
@@ -203,71 +148,11 @@ func ResolveProcessInstanceIncidentsOperation(ctx context.Context, api API, key 
 	return result, errorFromResult(result.Error)
 }
 
-func ResolveIncidents(ctx context.Context, api API, keys typex.Keys, wantedWorkers int, opts ...services.CallOption) (d.IncidentResolutionResults, error) {
-	cfg := services.ApplyCallOptions(opts)
-	ukeys := keys.Unique()
-	nw := toolx.DetermineNoOfWorkers(len(ukeys), wantedWorkers, cfg.NoWorkerLimit)
-	items, err := pool.ExecuteSlice[string, d.IncidentResolutionResult](ctx, ukeys, nw, cfg.FailFast, func(ctx context.Context, key string, _ int) (d.IncidentResolutionResult, error) {
-		return ResolveIncidentOperation(ctx, api, key, opts...)
-	})
-	items = compactIncidentResolutionResults(items)
-	return summarizeIncidentResolutionResults(items), err
-}
-
-func ResolveProcessInstancesIncidents(ctx context.Context, api API, keys typex.Keys, wantedWorkers int, opts ...services.CallOption) (d.ProcessInstanceResolutionResults, error) {
-	cfg := services.ApplyCallOptions(opts)
-	ukeys := keys.Unique()
-	nw := toolx.DetermineNoOfWorkers(len(ukeys), wantedWorkers, cfg.NoWorkerLimit)
-	items, err := pool.ExecuteSlice[string, d.ProcessInstanceResolutionResult](ctx, ukeys, nw, cfg.FailFast, func(ctx context.Context, key string, _ int) (d.ProcessInstanceResolutionResult, error) {
-		return ResolveProcessInstanceIncidentsOperation(ctx, api, key, opts...)
-	})
-	items = compactProcessInstanceResolutionResults(items)
-	return summarizeProcessInstanceResolutionResults(items), err
-}
-
 func mutationNotAcceptedMessage(status string) string {
 	if status != "" {
 		return status
 	}
 	return "incident resolution mutation was not accepted"
-}
-
-func summarizeIncidentResolutionResults(items []d.IncidentResolutionResult) d.IncidentResolutionResults {
-	out := d.IncidentResolutionResults{Operation: d.ResolutionOperationIncident, Items: items, Total: len(items)}
-	for _, item := range items {
-		out.DryRun = out.DryRun || item.DryRun
-		out.MutationSubmitted = out.MutationSubmitted || item.MutationSubmitted
-		switch item.Status {
-		case d.IncidentResolutionStatusSubmitted:
-			out.Submitted++
-		case d.IncidentResolutionStatusConfirmed:
-			out.Confirmed++
-		case d.IncidentResolutionStatusSkipped, d.IncidentResolutionStatusPlanned:
-			out.Skipped++
-		case d.IncidentResolutionStatusMutationFailed, d.IncidentResolutionStatusConfirmationFailed:
-			out.Failed++
-		}
-	}
-	return out
-}
-
-func summarizeProcessInstanceResolutionResults(items []d.ProcessInstanceResolutionResult) d.ProcessInstanceResolutionResults {
-	out := d.ProcessInstanceResolutionResults{Operation: d.ResolutionOperationProcessInstance, Items: items, Total: len(items)}
-	for _, item := range items {
-		out.DryRun = out.DryRun || item.DryRun
-		out.MutationSubmitted = out.MutationSubmitted || item.MutationSubmitted
-		switch item.Status {
-		case d.ProcessInstanceResolutionStatusSubmitted:
-			out.Submitted++
-		case d.ProcessInstanceResolutionStatusConfirmed:
-			out.Confirmed++
-		case d.ProcessInstanceResolutionStatusSkipped, d.ProcessInstanceResolutionStatusPlanned:
-			out.Skipped++
-		case d.ProcessInstanceResolutionStatusFailed, d.ProcessInstanceResolutionStatusPartialFailed:
-			out.Failed++
-		}
-	}
-	return out
 }
 
 func domainIncidentsForProcessInstance(key string, incidents []d.ProcessInstanceIncidentDetail) []d.ProcessInstanceIncidentDetail {
@@ -328,26 +213,4 @@ func errorFromResult(message string) error {
 		return nil
 	}
 	return errors.New(message)
-}
-
-func compactIncidentResolutionResults(items []d.IncidentResolutionResult) []d.IncidentResolutionResult {
-	out := items[:0]
-	for _, item := range items {
-		if item.IncidentKey == "" && item.Status == "" {
-			continue
-		}
-		out = append(out, item)
-	}
-	return out
-}
-
-func compactProcessInstanceResolutionResults(items []d.ProcessInstanceResolutionResult) []d.ProcessInstanceResolutionResult {
-	out := items[:0]
-	for _, item := range items {
-		if item.ProcessInstanceKey == "" && item.Status == "" {
-			continue
-		}
-		out = append(out, item)
-	}
-	return out
 }
